Update all.yml
Hi @hexylena, i would get the message ```condor_submit failed``` on some jobs on the new galaxy server.
I thought i would be becouse the new servers cant communicate with the old servers. thus we need to change the host and allow write. 

if this is not required or later please let me know.